Add config parameter fs.s3a.ssl.channel.mode=default_jsse_with_gcm.
The default value default_jsse would remove GSM-related cipher suites
during SSL handshake, causing handshake error. The idea of the property
apparently is to speed up the handshake because GSM cipher is slower.

The config parameter was not necessary earlier. So most likely, whether
Hadoop removes cipher suites is dependent on the performance of the
computer hosting the Hadoop processes.